// server/enums/notHopingToGetWorkValues.ts
enum NotHopingToGetWorkValues {
  FEELS_OFFENCE_WILL_LIMIT_ABILITY_TO_WORK = 'FEELS_OFFENCE_WILL_LIMIT_ABILITY_TO_WORK',
  HAS_FULL_TIME_CARING_RESPONSIBILITIES = 'HAS_FULL_TIME_CARING_RESPONSIBILITIES',
  LACKS_CONFIDENCE_OR_MOTIVATION = 'LACKS_CONFIDENCE_OR_MOTIVATION',
  NOT_ABLE_TO_WORK_DUE_TO_ILL_HEALTH = 'NOT_ABLE_TO_WORK_DUE_TO_ILL_HEALTH',
  REFUSED_SUPPORT_WITH_NO_REASON = 'REFUSED_SUPPORT_WITH_NO_REASON',
  RETIRED = 'RETIRED',
  RIGHT_TO_WORK_IN_UK_NOT_CONFIRMED = 'RIGHT_TO_WORK_IN_UK_NOT_CONFIRMED',
  OTHER = 'OTHER',
  NOT_SURE = 'NOT_SURE',
}

export default NotHopingToGetWorkValues